KEYCLOAK-6909 / KEYCLOAK-6986 Upstream product build stuff
This adds the build POM and assembly XMLs from the downstream repo, with some
improvements to the POM. Notably, jsonpath-maven-plugin and PME replace the
overly hacky ant string replacement logic for package.json

As this is only intended for building the product, there aren't community and
product profiles.

Since the product build pom depends on mvn properties, I have included
prod-arguments.json from KEYCLOAK-6909 in this same commit. This will allow a
build pipeline to substitute the appropriate values for things like the NPM
registry.

The community build is unaffected by this change. All the new files are kept in
a seperate product/ directory.