Add configurable host and port for Docker Compose
Using keycloak-connect within a Docker service is currently tricky. The
module uses the same host and port for both the auth server as well as
the app server. However, with docker-compose it is extremely easy to
have the two running separate containers, causing issues with what
exactly does localhost refer to, and also not properly mapping back to
the appropriate port number in the other container. This commit simply
adds the ability to specify these such that when a grant is obtained
from code, it the host and port to connect to inside the http request
can be customized.